DockerFR/requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
numpy
opencv-python-headless
Pillow
insightface
onnxruntime

